pyproject.toml
[project]
name = "vastai-connect"
version = "0.1.0"
description = "CLI wrapper for vastai - simplifies GPU rental workflow"
readme = "README.md"
authors = [
    { name = "Yusuke710", email = "youfu1202mo@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0.3",
    "questionary>=2.1.1",
    "vastai>=0.5.0",
]

[project.scripts]
vastai-connect = "vastai_connect.cli:main"

[build-system]
requires = ["uv_build>=0.8.0,<0.9"]
build-backend = "uv_build"
